# openapi.yml
openapi: 3.0.0
info:
  title: sampleapi
  description: 開発のサンプルです
  version: 1.0.0
paths:
  /articles:
    get:
      summary: Get all article
      operationId: getArticles
      description: Returns an array of articles
      responses:
        '200':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  result:
                    type: string
                  articles:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          minimum: 1
                        content:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                  message:
                    type: string
    parameters: []
    post:
      summary: Post Articles
      operationId: postArticles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                articles:
                  type: array
                  items:
                    type: object
                    properties:
                      content:
                        type: string
              required:
                - articles
        description: ''